Add sorted tree processor
Based off the other tree processor, so some code could be consolidated in a helper. I could have adjusted things a bit and done an inheritance thing here, but not too sure what is to be gained by that. Or could add sort support to the interface and the front to back processor would need something magical to make that happen.