Add validators
RegEx one seems too complex for nested scenarios and not a good fit. Implemented a recursive one treating the string like a function call with parameters that themselves could be function calls. Not yet implemented in processors, but will have them be composed of a validator.

Also had idea to use a stack for validation, which led me to just using a stack as a processor But musing on pushing and popping when coming across parentheses, I had a thought to use parentheses to denote a counter and associate that with whatever name I'm on and ignore whitespace and commas and display based off that.

Going to try that real quick before I write tests for the validator and inject it in processors.